FROMGIT: media: rc: add keymap for Khadas VIM/EDGE remote
This is a simple NEC remote control device shipped with the Khadas VIM
and EDGE range of ARM Single Board Computer (SBC) devices. It includes
a mouse button for Android use which is mapped to KEY_MUTE. This is the
sole difference from the Khadas 'geekbox' remote which shipped with an
earlier generation of Shenzen Wesion (Khadas) devices under the GeekBox
brand.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>