dt-bindings: arm: amlogic: switch X96 Max to use 100/1000 variants
X96 Max devices are also shipping in a cost-engineered 10/100 Ethernet
variant using internal RMII PHY and a lower-spec Ampak module. Change
to a common dtsi and spawn new dts files with -100 and -1000 names to
follow the established device-tree naming conventions used with popular
Amlogic legacy kernel distributions. The breaking change mainly impacts
users of Armbian and LibreELEC, but the disruption is manageable and it
is better for both to support an easier experience for future installs.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>