FROMGIT: ARM: dts: imx6-sr-som: increase at8035 PHY gigabit Tw parameter
Increase the SmartEEE Tw parameter for Atheros PHYs to stop gigabit
links from sporadically dropping. Jon Nettleton found that a value
of 23 was the minimum to give a stable link, but testing with the
Com Express 7 module shows 24 is the minimum - so 23 may still be on
the margins. Use 24 instead for consistency across SolidRun platforms.

Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>