WIP: media: dvb-frontends: add Sony CXD2841ER (WeTek) driver
Import the CXD2841ER demod driver from WeTek sources. This is used with the Ascot3
and MxL603 tuner drivers to support the DVB-C/T/T2 board. The upstream kernel has
an existing CXD2841ER driver that looks more maintained. We need to understand the
differences between them and have a strategy to abstract these through the use of
platform data in the driver linked to implementation comptable data in device-tree.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>