WIP: media: dvb-frontends: add WeTek NIMDETECT driver
Import the NIMDETECT driver from WeTek sources. This finds the i2s devices and
attempts to init them in an if-else sequence. Either this driver can/should be
cleaned up to follow the conventions in the dvb-frontend dvb_dummy driver, or
the content should be ported to the 'wetek' front-end experiment driver, or we
should create different front-ends for the three demod/tuner combos and create
three different board dts files to describe them.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>