FROMGIT(6.11): arm64: dts: meson: add GXLX/S905L/p271 support
Add a device-tree for the GXLX Amlogic P271 (S905L) reference design
board. This is a low-cost design similar to P281 (S905W) but with
silicon differences to omit VP9 and use Mali 450-MP2 (not MP3). The
SoC is marked with S905L and "2" (believed to denote MP2) resulting
in chip distributor stock lists and Android STB marketing sometimes
describing it as an S905L2 chip.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>