arm64: dts: meson: add p271 support
Add a device-tree for the GXLX Amlogic P271 (S905L) reference design board. This
is a low-cost design similar to P281 (S905W) and P212 (S905X) but with silicon
differences to omit VP9 and use Mali 450-MP2 (not MP3). The SoC is marked with
S905L and "2" (believed to denote MP2) resulting in some chip distributor stock
lists (and subsequent box vendor marketing) describing it as an S905L2 chip.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>