iio: adc: meson: add support for the GXLX SoC
The SARADC IP on GXLX is identical to the one found on GXL SoCs: except GXLX
requires poking the first three bits in the MESON_SAR_ADC_REG12 register to
get the three MPLL clocks (used as clock generators for the audio frequencies)
to work. Register values are taken from the vendor kernel.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>