FROMGIT(6.11): arm64: dts: amlogic: meson-g12b-bananapi: remove invalid fan on wrong pwm_cd controller
The fan is connected to the GPIOAO_10 pin, which can be connected to the pwm_AO_cd
controller, but this one is already used for the vddcpu_b regulator.

The fan was wrongly described as using the pwm_cd with the invalid "pwm-gpios"
property, both are wrong so remove the fan until we find an acceptable solution.

This fixes the following error:
pwm@1a000: 'pwm-gpios' does not match any of the regexes: 'pinctrl-[0-9]+'
        from schema $id: http://devicetree.org/schemas/pwm/pwm-amlogic.yaml#

Signed-off-by: Neil Armstrong <neil.armstrong@linaro.org>